refactor preprocessing
