+++ fedot/core/pipelines/pipeline.py
@@ -11,14 +11,13 @@ from fedot.core.dag.graph import Graph
 from fedot.core.data.data import InputData, data_has_categorical_features, data_has_missing_values, data_type_is_table
 from fedot.core.data.multi_modal import MultiModalData
 from fedot.core.log import Log, default_log
-from fedot.core.operations.evaluation.operation_implementations.data_operations.sklearn_transformations import \
-    DataOperationImplementation, ImputationImplementation, OneHotEncodingImplementation
 from fedot.core.optimisers.timer import Timer
 from fedot.core.optimisers.utils.population_utils import input_data_characteristics
 from fedot.core.pipelines.node import Node, PrimaryNode
+from fedot.core.pipelines.preprocessing import imputation_implementation, encode_data_for_prediction, \
+    encode_data_for_fit, pipeline_encoders_validation, custom_preprocessing
 from fedot.core.pipelines.template import PipelineTemplate
 from fedot.core.pipelines.tuning.unified import PipelineTuner
-
 
 # The allowed empirical partition limit of the number of rows to delete.
 # Rows that have 'string' type, instead of other 'integer' observes.
@@ -184,18 +183,6 @@ class Pipeline(Graph):
                                                         time=time_constraint)
         return train_predicted
 
-    def _preprocessing_fit_data(self, data: Union[InputData, MultiModalData]):
-        has_imputation_operation, has_encoder_operation = pipeline_encoders_validation(self)
-
-        data = _custom_preprocessing(data)
-
-        if data_has_missing_values(data) and not has_imputation_operation:
-            data = _imputation_implementation(data)
-
-        if data_has_categorical_features(data) and not has_encoder_operation:
-            self.pre_proc_encoders = _encode_data_for_fit(data)
-        return data
-
     @property
     def is_fitted(self):
         return all([(node.fitted_operation is not None) for node in self.nodes])
@@ -235,20 +222,38 @@ class Pipeline(Graph):
 
         # Make copy of the input data to avoid performing inplace operations
         copied_input_data = copy(input_data)
-        has_imputation_operation, has_encoder_operation = pipeline_encoders_validation(self)
-
-        copied_input_data = _custom_preprocessing(copied_input_data)
-
-        if data_has_missing_values(copied_input_data) and not has_imputation_operation:
-            copied_input_data = _imputation_implementation(copied_input_data)
-
-        if data_has_categorical_features(copied_input_data) and not has_encoder_operation:
-            _encode_data_for_prediction(copied_input_data, self.pre_proc_encoders)
-
+        copied_input_data = self._preprocessing_predict_data(copied_input_data)
         copied_input_data = self._assign_data_to_nodes(copied_input_data)
 
         result = self.root_node.predict(input_data=copied_input_data, output_mode=output_mode)
         return result
+
+    def _preprocessing_fit_data(self, data: Union[InputData, MultiModalData]):
+        """ Delete missing values from InputData for fitting """
+        has_imputation_operation, has_encoder_operation = pipeline_encoders_validation(self)
+
+        data = custom_preprocessing(data)
+
+        if data_has_missing_values(data) and not has_imputation_operation:
+            data = imputation_implementation(data)
+
+        if data_has_categorical_features(data) and not has_encoder_operation:
+            self.pre_proc_encoders = encode_data_for_fit(data)
+        return data
+
+    def _preprocessing_predict_data(self, data: Union[InputData, MultiModalData]):
+        """ Delete missing values from InputData for predict """
+        has_imputation_operation, has_encoder_operation = pipeline_encoders_validation(self)
+
+        copied_input_data = custom_preprocessing(data)
+
+        if data_has_missing_values(copied_input_data) and not has_imputation_operation:
+            copied_input_data = imputation_implementation(copied_input_data)
+
+        if data_has_categorical_features(copied_input_data) and not has_encoder_operation:
+            encode_data_for_prediction(copied_input_data, self.pre_proc_encoders)
+
+        return data
 
     def fine_tune_all_nodes(self, loss_function: Callable,
                             loss_params: dict = None,
@@ -343,120 +348,6 @@ class Pipeline(Graph):
             print(f"{node.operation.operation_type} - {node.custom_params}")
 
 
-def pipeline_encoders_validation(pipeline: Pipeline) -> (bool, bool):
-    """ Check whether Imputation and OneHotEncoder operation exist in pipeline.
-
-    :param data Pipeline: data to check
-    :return (bool, bool): has Imputation and OneHotEncoder in pipeline
-    """
-
-    has_imputers, has_encoders = [], []
-
-    def _check_imputer_encoder_recursion(root: Optional[Node], has_imputer: bool = False, has_encoder: bool = False):
-        node_type = root.operation.operation_type
-        if node_type == 'simple_imputation':
-            has_imputer = True
-        if node_type == 'one_hot_encoding':
-            has_encoder = True
-
-        if has_imputer and has_encoder:
-            return has_imputer, has_encoder
-        elif root.nodes_from is None:
-            return has_imputer, has_encoder
-
-        for node in root.nodes_from:
-            answer = _check_imputer_encoder_recursion(node, has_imputer, has_encoder)
-            if answer is not None:
-                imputer, encoder = answer
-                has_imputers.append(imputer)
-                has_encoders.append(encoder)
-
-    _check_imputer_encoder_recursion(pipeline.root_node)
-
-    if not has_imputers and not has_encoders:
-        return False, False
-
-    has_imputer = len([_ for _ in has_imputers if not _]) == 0
-    has_encoder = len([_ for _ in has_encoders if not _]) == 0
-    return has_imputer, has_encoder
-
-
-def _is_numeric(s) -> bool:
-    """ Check if variable converted to float.
-
-    :param s: any type variable
-    :return: is variable convertable to float
-    """
-    try:
-        float(s)
-        return True
-    except ValueError:
-        return False
-
-
-def is_np_array_has_nan(array):
-    for x in array:
-        if x is np.nan:
-            return True
-    return False
-
-
-def _try_convert_to_numeric(values):
-    try:
-        values = pd.to_numeric(values)
-        values = values.astype(np.number)
-    except ValueError:
-        pass
-    return values
-
-
-def _custom_preprocessing(data: Union[InputData, MultiModalData]):
-    if isinstance(data, InputData):
-        if data_type_is_table(data):
-            data = _preprocessing_input_data(data)
-    elif isinstance(data, MultiModalData):
-        for data_source_name, values in data.items():
-            if data_type_is_table(values):
-                data[data_source_name] = _preprocessing_input_data(values)
-
-    return data
-
-
-def _preprocessing_input_data(data: InputData) -> InputData:
-    features = data.features
-    target = data.target
-
-    # delete rows with equal target None
-    if target is not None and len(target.shape) != 0 and is_np_array_has_nan(target):
-        target_index_with_nan = np.hstack(np.argwhere(np.isnan(target)))
-        data.features = np.delete(features, target_index_with_nan, 0)
-        data.target = np.delete(data.target, target_index_with_nan, 0)
-        data.idx = np.delete(data.idx, target_index_with_nan, 0)
-
-    source_shape = features.shape
-    columns_amount = source_shape[1] if len(source_shape) > 1 else 1
-
-    for i in range(columns_amount):
-        values = pd.Series(features[:, i])
-        # check for each column, if values converted to numeric. remember index of rows that not converted
-        if any(list(map(lambda x: isinstance(x, str), values))):
-            not_numeric = list(map(lambda x: not _is_numeric(x), values))
-            rows_to_nan = list(values.index[not_numeric])
-            partition_not_numeric = len(rows_to_nan) / source_shape[0]
-
-            # if partition of numerical rows less then EMPIRICAL_PARTITION,
-            # then convert to numerical and others to Nan
-            if partition_not_numeric < EMPIRICAL_PARTITION:
-                values[rows_to_nan] = np.nan
-                data.features[:, i] = _try_convert_to_numeric(values)
-            # if EMPIRICAL_PARTITION < partition < 1, then some data in column are
-            # integer and some data are string, can not handle this case
-            elif partition_not_numeric < 0.9:
-                raise ValueError("The data in the column has a different type. Need to preprocessing data manually.")
-
-    return data
-
-
 def nodes_with_operation(pipeline: Pipeline, operation_name: str) -> list:
     """ The function return list with nodes with the needed operation
 
@@ -469,87 +360,3 @@ def nodes_with_operation(pipeline: Pipeline, operation_name: str) -> list:
     appropriate_nodes = filter(lambda x: x.operation.operation_type == operation_name, pipeline.nodes)
 
     return list(appropriate_nodes)
-
-
-def _encode_data_for_fit(data: Union[InputData, MultiModalData]) -> \
-        Union[List[DataOperationImplementation], DataOperationImplementation]:
-    """ Encode categorical features to numerical. In additional,
-    save encoders to use later for prediction data.
-
-    :param data: data to transform
-    :return encoders: operation preprocessing categorical features or list of it
-    """
-
-    encoders = None
-    if isinstance(data, InputData):
-        transformed, encoder = _create_encoder(data)
-        encoders = encoder
-        data.features = transformed
-    elif isinstance(data, MultiModalData):
-        encoders = {}
-        for data_source_name, values in data.items():
-            if data_source_name.startswith('data_source_table'):
-                transformed, encoder = _create_encoder(values)
-                if encoder is not None:
-                    encoders[data_source_name] = encoder
-                data[data_source_name].features = transformed
-
-    return encoders
-
-
-def _encode_data_for_prediction(data: Union[InputData, MultiModalData],
-                                encoders: Union[dict, DataOperationImplementation]):
-    """ Transformation the prediction data inplace. Use the same transformations as for the training data.
-
-    :param data: data to transformation
-    :param encoders: encoders f transformation
-    """
-    if encoders:
-        if isinstance(data, InputData):
-            transformed = encoders.transform(data, True).predict
-            data.features = transformed
-        elif isinstance(data, MultiModalData):
-            for data_source_name, encoder in encoders.items():
-                transformed = encoder.transform(data[data_source_name], True).predict
-                data[data_source_name].features = transformed
-
-
-def _imputation_implementation(data: Union[InputData, MultiModalData]) -> Union[InputData, MultiModalData]:
-    if isinstance(data, InputData):
-        return _imputation_implementation_unidata(data)
-    if isinstance(data, MultiModalData):
-        for data_source_name, values in data.items():
-            if data_source_name.startswith('data_source_table') or data_source_name.startswith('data_source_ts'):
-                data[data_source_name].features = _imputation_implementation_unidata(values)
-        return data
-    raise ValueError(f"Data format is not supported.")
-
-
-def _imputation_implementation_unidata(data: InputData):
-    """ Fill in the gaps in the data inplace.
-
-    :param data: data for fill in the gaps
-    """
-    imputer = ImputationImplementation()
-    output_data = imputer.fit_transform(data)
-    transformed = InputData(features=output_data.predict, data_type=output_data.data_type,
-                            target=output_data.target, task=output_data.task, idx=output_data.idx)
-    return transformed
-
-
-def _create_encoder(data: InputData):
-    """ Fills in the gaps, converts categorical features using OneHotEncoder and create encoder.
-
-    :param data: data to preprocess
-    :return tuple(array, Union[OneHotEncodingImplementation, None]): tuple of transformed and [encoder or None]
-    """
-
-    encoder = None
-    if data_has_categorical_features(data):
-        encoder = OneHotEncodingImplementation()
-        encoder.fit(data)
-        transformed = encoder.transform(data, True).predict
-    else:
-        transformed = data.features
-
-    return transformed, encoder
