+++ test/unit/pipelines/test_pipeline.py
@@ -15,7 +15,8 @@ from sklearn.metrics import roc_auc_score as roc
 from fedot.core.data.data import InputData
 from fedot.core.data.data_split import train_test_data_setup
 from fedot.core.pipelines.node import PrimaryNode, SecondaryNode
-from fedot.core.pipelines.pipeline import Pipeline, pipeline_encoders_validation
+from fedot.core.pipelines.pipeline import Pipeline
+from fedot.core.pipelines.preprocessing import pipeline_encoders_validation
 from fedot.core.repository.dataset_types import DataTypesEnum
 from fedot.core.repository.tasks import Task, TaskTypesEnum, TsForecastingParams
 from fedot.core.utils import probs_to_labels
